Utilize source field for confidence and answered
